refactor: drop QM ports and devices
