+++ src/qibolab/instruments/qm/__init__.py
@@ -1,2 +1,2 @@
 from .controller import QMController
-from .devices import Octave, OPXplus
+from .octave import Octave
